+++ kong/response.go
@@ -34,6 +34,21 @@ func hasError(res *http.Response) error {
 		return nil
 	}
 
-	body, _ := io.ReadAll(res.Body) // TODO error in error?
+	body, err := io.ReadAll(res.Body)
+	if err != nil {
+		return fmt.Errorf("failed to read error body: %w", err)
+	}
 	return NewAPIError(res.StatusCode, messageFromBody(body))
+}
+
+func hasErrorRaw(res *http.Response) error {
+	if res.StatusCode >= 200 && res.StatusCode <= 399 {
+		return nil
+	}
+
+	body, err := io.ReadAll(res.Body)
+	if err != nil {
+		return fmt.Errorf("failed to read error body: %w", err)
+	}
+	return NewAPIErrorWithRaw(res.StatusCode, messageFromBody(body), body)
 }